Refactoring of remote switch classes.

+++ SDK/Hardware/CK.HomeAutomation.Hardware.RemoteSwitch/LPD433MhzCode.cs
@@ -2,13 +2,13 @@
 {
     public class LPD433MhzCode
     {
-        public LPD433MhzCode(ulong value, int length)
+        public LPD433MhzCode(uint value, int length)
         {
             Value = value;
             Length = length;
         }
 
-        public ulong Value { get; }
+        public uint Value { get; }
 
         public int Length { get; }
     }
